Clear some TODOs
The TODO at line 176 was actually already fixed but an and_eq is slightly simpler than the previous statements.
For the TODO at line 220: whether or not this is _needed_ is very difficult to determine. Sometimes NC decides not post a notification because the application posting it is already 'active' which doesn't really apply to terminal-notifier. However, having this here ensure that the notification is posted either way, so it probably good to have.