Only prepend file:// scheme if there is no scheme
'appIcon' and 'contentImage' options could actually accept data:,
http(s):, etc URIs. However, the fact that the current version checks
explicitly for file:// URIs cripples that use case.

This commit tries to treat the given string as a URI. However, if the
URI does not have a scheme, then 'file://' is appended and the resulting
string is re-interpreted as the final URI. Otherwise, we are set anyway.